WIP: #2 First TS classes, work stopped because of the necessary fixes in the `ExcelJS` library: https://github.com/exceljs/exceljs/pull/1092
